Feature/plmc 229 implement new vesting pallet on plmc bond

+++ pallets/funding/src/mock.rs
@@ -270,7 +270,7 @@ impl pallet_funding::Config for TestRuntime {
 	type Randomness = RandomnessCollectiveFlip;
 	type RemainderFundingDuration = RemainderFundingDuration;
 	type RuntimeEvent = RuntimeEvent;
-	type StorageItemId = u128;
+	type StorageItemId = u32;
 	type StringLimit = ConstU32<64>;
 	type SuccessToSettlementTime = SuccessToSettlementTime;
 	type TreasuryAccount = TreasuryAccount;
